Blocks, Ores, and Ingots!
This is merely the first update to the vast amount of ores planned, and
sorry, no metadata blocks just yet (Grr, GameRegistry.addSmelting)...
Planned for the rewrite of MachiniTech, meaning that this mod is mere at
this point for us to get to learn the ins and outs of Minecraft Forge
and GitHub.

On another note, you may notice that MachiniTechOreStorageBlock does not
extend net.minecraft.block.BlockOreStorage. This is merely a choice of
convenience, although to make this mod a better example to others, I may
change it in the future.

-I don't see the money. Where is it? (Hahaha,very funny)