Ready for a first release!
Still got a ways to go, but we're getting there! Still have to write the
multiblock API, and learn about tile entities, but for now, I'd say that
this deserves to System.out.println("Hello, World!");

And yes, I realise I missed a "static" in the method last time. Please
don't hurt me.