Tools, Tile Entities and More!
It's been a while, but we're back with tools and a configuration file.
In the background, the Smelter is taking shape, although it's not quite
done yet. Then, there's plans for the multiblock smelter, but that's for
another day. Gotta finish the smaller, single block smelter first.

And if you're wondering, textures exist only for copper tools because
I'm lazy. Feel free to contribute some textures.