Tweaks and a faster hopper
Smelter's still not where it should be, still a glorified furnace, but
will change later.

And I'm fed up with the slow vanilla hopper. So I made a new one. It's
99% Mojang code, there's only a few changes, but it goes much faster.
Can be disabled in the config.

-Added heat mechanic to smelter, having an item burning will increase
heat to a maximum of 2000 without a coil
-Smelter smelts faster with more heat
-Removed ability of heating coils to increase burn time of item in
smelter
-Added heat field to Smelter Gui & Container
-Added addon manager to add onto vanilla and other mods
-Added faster hopper, can be disabled configs, 8 times faster than
vanilla hoppers
-Added support for ore storage blocks used in beacons

Still waiting on a recommended build for 1.6.1